requirements.txt
playwright
pandas
pyyaml
openai
reportlab
PyPDF2
